refactor(lab): keep run failure extraction in core
